+++ Packages/com.unity.inputsystem/InputSystem/Devices/InputDevice.cs
@@ -524,14 +524,23 @@ namespace UnityEngine.InputSystem
             where TCommand : struct, IInputDeviceCommandInfo
         {
             var commandPtr = (InputDeviceCommand*)UnsafeUtility.AddressOf(ref command);
+
             // Give callbacks first shot.
             var manager = InputSystem.s_Manager;
-            var callbacks = manager.m_DeviceCommandCallbacks;
+            var callbacks = manager.m_DeviceCommandCallbacks.PrepareExecution();
             for (var i = 0; i < callbacks.length; ++i)
             {
-                var result = callbacks[i](this, commandPtr);
-                if (result.HasValue)
-                    return result.Value;
+                try
+                {
+                    var result = callbacks[i](this, commandPtr);
+                    if (result.HasValue)
+                        return result.Value;
+                }
+                catch (Exception exception)
+                {
+                    Debug.LogError($"{exception.GetType().Name} while executing 'InputSystem.onDeviceCommand' callbacks");
+                    Debug.LogException(exception);
+                }
             }
 
             return ExecuteCommand((InputDeviceCommand*)UnsafeUtility.AddressOf(ref command));