+++ Packages/com.unity.inputsystem/InputSystem/InputSystem.cs
@@ -2769,16 +2769,13 @@ namespace UnityEngine.InputSystem
             {
                 if (value == null)
                     throw new ArgumentNullException(nameof(value));
-                lock (s_Manager)
-                    if (!InputActionState.s_OnActionChange.Contains(value))
-                        InputActionState.s_OnActionChange.Append(value);
+                InputActionState.s_OnActionChange.AddCallback(value);
             }
             remove
             {
                 if (value == null)
                     throw new ArgumentNullException(nameof(value));
-                lock (s_Manager)
-                    InputActionState.s_OnActionChange.Remove(value);
+                InputActionState.s_OnActionChange.RemoveCallback(value);
             }
         }